requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
chromadb
pymupdf
pytesseract
Pillow
openai-whisper
requests
python-multipart
langdetect
python-docx
